Restructure attribute bindings module
