+++ src/attribute/bindings/mod.rs
@@ -1,0 +1,4 @@
+mod numeric;
+mod string;
+
+pub(super) use string::*;
